유저가 받는거 까지 완료
